Synchro

+++ app/models/Bookmark.scala
@@ -7,13 +7,13 @@ package models
  * Time: 00:34
  * To change this template use File | Settings | File Templates.
  */
-case class Bookmark(id: Option[Long], title: String, url: String, details: String, category: Option[Category])
+case class Bookmark(id: Option[Int], title: String, url: String, details: String, category: Option[Category])
 
   object Bookmark {
 
     def all(): List[Bookmark] = Nil
 
-    def delete(id: Long) {}
+    def delete(id: Int) {}
 
     def save(bookmark: Bookmark) : Bookmark = {
       //TODO
